infra/exchanges/bybit/futures_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests==2.32.3
pytest==8.3.3
pydantic-settings==2.5.2
PyYAML==6.0.2
